Renamed files for `JoltCustomRayShape`
