+++ src/servers/jolt_globals.cpp
@@ -1,8 +1,8 @@
 #include "jolt_globals.hpp"
 
 #include "shapes/jolt_custom_empty_shape.hpp"
+#include "shapes/jolt_custom_ray_shape.hpp"
 #include "shapes/jolt_custom_user_data_shape.hpp"
-#include "shapes/jolt_ray_shape.hpp"
 
 void* jolt_alloc(size_t p_size) {
 	return mi_malloc(p_size);
