+++ src/shapes/jolt_custom_ray_shape.hpp
